backend/requirements.txt
fastapi
uvicorn
httpx
sqlalchemy
psycopg2-binary
supabase
python-multipart
python-dotenv
websockets
